created helper.js with all the common functions and constants (urls), created test case 20, finished sign-up page testing
